scripts/aiops/requirements.txt
# Core dependencies
neo4j>=5.14.0
pyyaml>=6.0.1

# Kafka
kafka-python>=2.0.2

# Redis
redis>=5.0.1

# HTTP requests
requests>=2.31.0

# Data processing
pandas>=2.1.4
numpy>=1.26.2

# Time series analysis
statsmodels>=0.14.0
scipy>=1.11.4

# Machine learning
scikit-learn>=1.3.2

# Anomaly detection
pyod>=1.1.3

# Graph algorithms
networkx>=3.2.1

# Logging and monitoring
structlog>=23.2.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
